Rename registry
